feat: replace AutoLoginMiddleware with explicit API login
Remove AutoLoginMiddleware and consolidate HistoryMiddleware into
middleware.py. Frontend now supports ?autologin query param for dev
convenience. Seed script logs in via API with proper error handling.
